+++ bot.py
@@ -1,4 +1,4 @@
-__version__ = "3.2.2"
+__version__ = "3.2.3-pre"
 
 import asyncio
 import logging